
\name{xem-package}
\alias{xem-package}
\alias{xem}
\docType{package}
\title{Support functions for the xeona energy systems modeling environment}
\description{Set of functions for recovering data from \emph{xeona} model (XEM) files and for subsequent manipulation.}
\details{
  \tabular{ll}{
    Package: \tab xem\cr
    Type: \tab package\cr
    Version: \tab 0.1\cr
    Date: \tab 2011-12-07\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
}}
\author{Robbie Morrison <robbie@actrix.co.nz>.}
%  \references{}
\keyword{package}

% end of file

